# Conflicts:
#	src/projects/base/common_types.h
#	src/projects/base/info/media_track.cpp
#	src/projects/base/info/media_track.h
#	src/projects/base/info/stream_info.h
#	src/projects/base/publisher/application.cpp
#	src/projects/config/config_manager.cpp
#	src/projects/media_router/media_route_application.cpp
#	src/projects/media_router/media_route_stream.cpp
#	src/projects/media_router/media_route_stream.h
#	src/projects/providers/rtmp/rtmp_chunk_stream.h
#	src/projects/providers/rtmp/rtmp_server.cpp
#	src/projects/publishers/webrtc/rtc_application.cpp
#	src/projects/relay/relay_client.cpp
#	src/projects/relay/relay_client.h
#	src/projects/relay/relay_datastructure.h
#	src/projects/relay/relay_server.cpp
#	src/projects/rtmp/rtmp_provider.h
#	src/projects/transcode/codec/transcode_codec_enc_avc.cpp
#	src/projects/transcode/transcode_stream.cpp
#	src/projects/transcode/transcode_stream.h

+++ src/projects/base/ovcrypto/base_64.cpp
@@ -40,7 +40,7 @@ namespace ov
 
 			if(result > 0)
 			{
-				OV_ASSERT2(result == data->GetLength());
+				OV_ASSERT2(static_cast<size_t>(result) == data->GetLength());
 				// 성공
 				break;
 			}
